refactor: 전체 레이아웃 구조 리팩토링 및 공통 컴포넌트 정리
